лаб5/requirements.txt
requests
numpy
pandas
matplotlib
beautifulsoup4
sqlalchemy
pytest
pyyaml
flask
pyttsx3